ecolar-mobile/src/main/webapp/app/entities/movement/movement.service.ts
import axios from 'axios';

import { IMovement } from '@/shared/model/movement.model';

const baseApiUrl = 'api/movements';

export default class MovementService {
    public find(id): Promise<IMovement> {
        return new Promise<IMovement>(resolve => {
            axios.get(`${baseApiUrl}/${id}`).then(function(res) {
                resolve(res.data);
            });
        });
    }

    public retrieve(paginationQuery?: any): Promise<any> {
        return new Promise<any>(resolve => {
            axios.get(baseApiUrl).then(function(res) {
                resolve(res);
            });
        });
    }

    public delete(id): Promise<any> {
        return new Promise<any>(resolve => {
            axios.delete(`${baseApiUrl}/${id}`).then(function(res) {
                resolve(res);
            });
        });
    }

    public create(entity): Promise<IMovement> {
        return new Promise<IMovement>(resolve => {
            axios.post(`${baseApiUrl}`, entity).then(function(res) {
                resolve(res.data);
            });
        });
    }

    public update(entity): Promise<IMovement> {
        return new Promise<IMovement>(resolve => {
            axios.put(`${baseApiUrl}`, entity).then(function(res) {
                resolve(res.data);
            });
        });
    }
}
